[Refactor] 고객 부분 파일 정리
