+++ src/App.js
@@ -4,7 +4,7 @@ import { Layout } from './layout/Layout';
 import { Login } from './pages/Login';
 import { Home } from './pages/Home';
 import { Signup } from './pages/Signup';
-import { CustomerKitchenWrite } from './pages/customer/mypage/CustomerKitchenWrite';
+import { CustomerKitchenWrite } from './pages/customer/CustomerKitchenWrite';
 import { ThemeProvider } from 'styled-components';
 import { theme } from './theme/theme';
 import Reserve from './pages/chef/Reserve';
@@ -17,8 +17,8 @@ import { Phone } from './pages/customer/mypage/informchange/Phone';
 import { Certification } from './pages/customer/mypage/informchange/Certification';
 import { CerDone } from './pages/customer/mypage/informchange/CerDone';
 import { CustomerBoard } from './pages/customer/CustomerBoard';
-import { HomeParty } from './pages/customer/customerboard/HomeParty';
-import { ChefDetail } from './pages/customer/customerboard/ChefDetail';
+import { HomeParty } from './modal/HomeParty';
+import { ChefDetail } from './modal/ChefDetail';
 import { CustomerHistory } from './pages/customer/CustomerHistory';
 import { ReviewPage } from './pages/customer/ReviewPage';
 import { ChefList } from './pages/customer/ChefList';
