+++ src/modal/RegistDone.jsx
